modify test to retrieve access token via login, seperate app listen from export of app with server/index files for jest
