+++ package.json
@@ -7,6 +7,7 @@
     "start": "node index.js",
     "server": "node index.js",
     "client": "cd client && npm start",
+    "test": "jest --runInBand --forceExit",
     "db:migrate": "cd db && npx sequelize-cli db:migrate", 
     "db:migrate:undo": "cd db && npx sequelize-cli db:migrate:undo:all", 
     "db:seed": "cd db && npx sequelize-cli db:seed:all",
